requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic[email]
websockets
passlib[bcrypt]
requests
bson
pytz